move options popup
