+++ client/src/pages/search/components/Options2.js
@@ -1,12 +1,12 @@
-import React, { useState, useEffect, useRef } from "react";
+import React from "react";
 import '../styles/Options2.css';
 
-const Options2 = ({ open, top, left }) => {
+const Options2 = ({ open, top, left, deleteOnClick }) => {
 
   return (
     <div className={`options-container ${open ? "active" : "inactive"}`} style={{ top: top, left: left }}>
         <div className="options-item">Edit</div>
-        <div className="options-item">Delete</div>
+        <div className="options-item" onClick={deleteOnClick}>Delete</div>
     </div>
   );
 }
